Работа над 3м этапом

+++ core/interaction-api/src/main/java/ru/yandex/practicum/DTO/event/EventShortDto.java
@@ -25,5 +25,5 @@ public class EventShortDto {
     Long initiator;
     Boolean paid;
     String title;
-    Long views;
+    Double views;
 }
